@@ -14,12 +14,11 @@ classifiers = [
   'Programming Language :: Python :: 3.11',
   'Programming Language :: Python :: 3.12',
   'Programming Language :: Python :: 3.13',
-  'Programming Language :: Python :: 3.9',
   'Topic :: Scientific/Engineering :: Information Analysis',
 ]
 dependencies = [
   'matplotlib>=3.0.1',
-  'numpy>=1.21.0', # minimum typing support
+  'numpy>=1.21.0',
   'pillow',
   'pooch',
   'scooby>=0.5.1',
@@ -27,6 +26,7 @@ dependencies = [
   'vtk!=9.4.0',
   'vtk!=9.4.1',
   'vtk<9.6.0',
+  'vtk>=9.2.2',
 ]
 description = 'Easier Pythonic interface to VTK'
 dynamic = ['version']
@@ -34,7 +34,7 @@ keywords = ['mesh', 'numpy', 'plotting', 'vtk']
 license = 'MIT'
 name = 'pyvista'
 readme = 'README.rst'
-requires-python = '>=3.9'
+requires-python = '>=3.10'
 
 [project.optional-dependencies]
 all = ['pyvista[colormaps,io,jupyter]']
@@ -85,6 +85,7 @@ test = [
   'numpydoc<1.10.0',
   'pandas<=2.4.0',
   'pyanalyze<=0.13.1',
+  'pyobjc-framework-Cocoa; sys_platform == "darwin"', # see pyvista/pulls/7828
   'pytest-cov<6.3.0',
   'pytest-mypy-plugins<3.3.0',
   'pytest-pyvista==0.2.0',
@@ -212,7 +213,6 @@ filterwarnings = [
   'error',
 
   "ignore:'mode' parameter is deprecated and will be removed in Pillow 13:DeprecationWarning", # matplotlib toolbar icon creation with newer Pillow
-  'ignore:(?s).*PYVISTA_VTK_DATA is not a valid directory.*:UserWarning',
   'ignore:.*A NumPy version .* is required for this version of SciPy.*:UserWarning',
   'ignore:.*Given trait value dtype "float64":UserWarning', # bogus numpy ABI warning (see numpy/#432)
   'ignore:.*The NumPy module was reloaded*:UserWarning', # bogus numpy ABI warning (see numpy/#432)
@@ -222,6 +222,7 @@ filterwarnings = [
   'ignore:The .*interactive_bk.* attribute was deprecated in Matplotlib 3\.9.*:matplotlib.MatplotlibDeprecationWarning', # https://github.com/microsoft/debugpy/issues/1623
   'ignore:pyvista test \w+ image dir.* does not yet exist.  Creating dir.:UserWarning',
 
+  'always:(?s).*PYVISTA_VTK_DATA is not a valid directory.*:UserWarning',
   'always:.*Exceeded image regression warning of .* with an image error of .*:UserWarning',
 ]
 
@@ -375,6 +376,7 @@ ignore = [
   'ANN401', # Disallow `Any`
   'B028', # https://github.com/pyvista/pyvista/pull/6030
   'B904', # https://github.com/pyvista/pyvista/pull/6022
+  'B905', # zip() without explicit strict parameter - will be addressed in separate PR
   'COM812', # May cause conflicts when used with the formatter
   'D105', # Missing docstring in magic method
   'D107', # Missing docstring in `__init__`
@@ -441,6 +443,8 @@ allow-dict-calls-with-keyword-arguments = true
 'pyvista/__init__.py' = ['ANN']
 'pyvista/_plot.py' = ['ANN']
 'pyvista/conftest.py' = ['ANN']
+'pyvista/core/_typing_core/_aliases.py' = ['PYI047', 'UP007']
+'pyvista/core/_typing_core/_array_like.py' = ['PYI047', 'UP007']
 'pyvista/core/_validation/_cast_array.py' = ['ANN']
 'pyvista/core/_validation/validate.py' = ['ANN202']
 'pyvista/core/_vtk_core.py' = ['ANN', 'PLC0414']
