@@ -18,7 +18,6 @@ from __future__ import annotations
 
 import time
 
-import pyvista as pv
 from pyvista import examples
 
 lidar = examples.download_lidar()
@@ -34,7 +33,6 @@ clipped.plot(show_scalar_bar=False)
 
 # %%
 # Show the performance improvement when using a PointSet.
-# This is only available with VTK >= 9.1.0.
 #
 # .. note::
 #   For VTK 9.4.X, there is no performance improvement due to a VTK bug which
@@ -42,12 +40,11 @@ clipped.plot(show_scalar_bar=False)
 
 # pset = lidar.cast_to_pointset()
 
-if pv.vtk_version_info >= (9, 1):
-    lidar_pset = lidar.cast_to_pointset()
-    tstart = time.time()
-    clipped_pset = lidar_pset.clip(origin=(0, 0, 1.76e3), normal=(0, 0, 1))
-    t_elapsed = time.time() - tstart
-    print(f'Time to clip with a PointSet {t_elapsed:.2f} seconds.')
+lidar_pset = lidar.cast_to_pointset()
+tstart = time.time()
+clipped_pset = lidar_pset.clip(origin=(0, 0, 1.76e3), normal=(0, 0, 1))
+t_elapsed = time.time() - tstart
+print(f'Time to clip with a PointSet {t_elapsed:.2f} seconds.')
 
 # %%
 # Plot the same dataset.
@@ -56,7 +53,6 @@ if pv.vtk_version_info >= (9, 1):
 #    PyVista must still create an intermediate PolyData to be able to plot, so
 #    there is no performance improvement when using a :class:`pyvista.PointSet`
 
-if pv.vtk_version_info >= (9, 1):
-    clipped_pset.plot(show_scalar_bar=False)
+clipped_pset.plot(show_scalar_bar=False)
 # %%
 # .. tags:: load
