@@ -28,6 +28,8 @@ from pyvista.core.wrappers import _wrappers as _wrappers
 from pyvista.jupyter import set_jupyter_backend as set_jupyter_backend
 from pyvista.report import GPUInfo as GPUInfo
 from pyvista.report import Report as Report
+from pyvista.report import check_math_text_support as check_math_text_support
+from pyvista.report import check_matplotlib_vtk_compatibility as check_matplotlib_vtk_compatibility
 from pyvista.report import get_gpu_info as get_gpu_info
 
 if TYPE_CHECKING:
@@ -36,11 +38,11 @@ if TYPE_CHECKING:
 # get the int type from vtk
 ID_TYPE: type[np.int32 | np.int64] = _get_vtk_id_type()
 
-# determine if using at least vtk 9.0.0
-if vtk_version_info.major < 9:  # pragma: no cover
+# determine if using at least vtk 9.1.0
+if vtk_version_info < (9, 2, 0):  # pragma: no cover
     from pyvista.core.errors import VTKVersionError
 
-    msg = 'VTK version must be 9.0.0 or greater.'
+    msg = 'VTK version must be 9.2.0 or greater.'
     raise VTKVersionError(msg)
 
 # catch annoying numpy/vtk future warning:
