@@ -39,17 +39,6 @@ def _lazy_vtk_instantiation(module_name, class_name):
     return getattr(module, class_name)()
 
 
-def lazy_vtkPOpenFOAMReader():
-    """Lazy import of the :vtk:`vtkPOpenFOAMReader`."""
-    from vtkmodules.vtkIOParallel import vtkPOpenFOAMReader  # noqa: PLC0415
-    from vtkmodules.vtkParallelCore import vtkDummyController  # noqa: PLC0415
-
-    # Workaround waiting for the fix to be upstream (MR 9195 gitlab.kitware.com/vtk/vtk)
-    reader = vtkPOpenFOAMReader()
-    reader.SetController(vtkDummyController())
-    return reader
-
-
 def get_reader(filename, force_ext=None):
     """Get a reader for fine-grained control of reading data files.
 
@@ -1232,9 +1221,8 @@ class POpenFOAMReader(OpenFOAMReader):
     parallel reconstructed data, and decomposed data.
     """
 
-    _class_reader = staticmethod(lazy_vtkPOpenFOAMReader)
-    _vtk_module_name = ''
-    _vtk_class_name = ''
+    _vtk_module_name = 'vtkIOParallel'
+    _vtk_class_name = 'vtkPOpenFOAMReader'
 
     @property
     def case_type(self):
@@ -1546,10 +1534,7 @@ class MultiBlockPlot3DReader(BaseReader):
 
         # AddFileName supports reading multiple q files
         for q_filename in files:
-            if pyvista.vtk_version_info < (9, 2, 2):  # pragma no cover
-                self.reader.AddFileName(str(q_filename))
-            else:
-                self.reader.AddFileName(q_filename)
+            self.reader.AddFileName(q_filename)
 
     @property
     def auto_detect_format(self):
@@ -2608,13 +2593,6 @@ class HDFReader(BaseReader):
 
     _vtk_module_name = 'vtkIOHDF'
     _vtk_class_name = 'vtkHDFReader'
-
-    def __init__(self, path):
-        if pyvista.vtk_version_info < (9, 1, 0):
-            msg = f'{self.__class__.__name__} is only available for vtk>=9.1'
-            raise pyvista.VTKVersionError(msg)
-
-        super().__init__(path)
 
     @wraps(BaseReader.read)
     def read(self):  # type: ignore[override]
