@@ -5,11 +5,9 @@ from __future__ import annotations
 from enum import Enum
 import os
 import platform
-import subprocess
 from subprocess import PIPE
 from subprocess import Popen
 from subprocess import TimeoutExpired
-import sys
 
 import numpy as np
 
@@ -725,60 +723,36 @@ def parse_font_family(font_family: str) -> int:
     return FONTS[font_family].value
 
 
-def check_matplotlib_vtk_compatibility():
-    """Check if VTK and Matplotlib versions are compatible for MathText rendering.
-
-    This function is primarily geared towards checking if MathText rendering is
-    supported with the given versions of VTK and Matplotlib. It follows the
-    version constraints:
-
-    * VTK <= 9.2.2 requires Matplotlib < 3.6
-    * VTK > 9.2.2 requires Matplotlib >= 3.6
-
-    Other version combinations of VTK and Matplotlib will work without
-    errors, but some features (like MathText/LaTeX rendering) may
-    silently fail.
+def check_math_text_support() -> bool:  # pragma: no cover
+    """Raise a DeprecationError as this has been moved.
 
     Returns
     -------
     bool
-        True if the versions of VTK and Matplotlib are compatible for MathText
-        rendering, False otherwise.
-
-    Raises
-    ------
-    RuntimeError
-        If the versions of VTK and Matplotlib cannot be checked.
+        Returns False for compatibility.
 
     """
-    import matplotlib as mpl  # noqa: PLC0415
+    from pyvista.core.errors import DeprecationError  # noqa: PLC0415
 
-    mpl_vers = tuple(map(int, mpl.__version__.split('.')[:2]))
-    if pyvista.vtk_version_info <= (9, 2, 2):
-        return not mpl_vers >= (3, 6)
-    elif pyvista.vtk_version_info > (9, 2, 2):
-        return mpl_vers >= (3, 6)
-    msg = 'Uncheckable versions.'  # pragma: no cover
-    raise RuntimeError(msg)  # pragma: no cover
+    # Deprecated on v0.47.0, estimated removal on v0.50.0
+    msg = '`check_math_text_support` is now imported from `pyvista.report`'
+    DeprecationError(msg)
+
+    return False
 
 
-def check_math_text_support():
-    """Check if MathText and LaTeX symbols are supported.
+def check_matplotlib_vtk_compatibility() -> bool:  # pragma: no cover
+    """Raise a DeprecationError as this has been moved.
 
     Returns
     -------
     bool
-        ``True`` if both MathText and LaTeX symbols are supported, ``False``
-        otherwise.
+        Returns False for compatibility.
 
     """
-    # Something seriously sketchy is happening with this VTK code
-    # It seems to hijack stdout and stderr?
-    # See https://github.com/pyvista/pyvista/issues/4732
-    # This is a hack to get around that by executing the code in a subprocess
-    # and capturing the output:
-    # _vtk.vtkMathTextFreeTypeTextRenderer().MathTextIsSupported()
-    _cmd = 'import vtk;print(vtk.vtkMathTextFreeTypeTextRenderer().MathTextIsSupported());'
-    proc = subprocess.run([sys.executable, '-c', _cmd], check=False, capture_output=True)
-    math_text_support = False if proc.returncode else proc.stdout.decode().strip() == 'True'
-    return math_text_support and check_matplotlib_vtk_compatibility()
+    from pyvista.core.errors import DeprecationError  # noqa: PLC0415
+
+    # Deprecated on v0.47.0, estimated removal on v0.50.0
+    msg = '`check_matplotlib_vtk_compatibility` is now imported from `pyvista.report`'
+    DeprecationError(msg)
+    return False  # returning bool for compatibility
