@@ -30,7 +30,8 @@ from tests.core.test_dataset_filters import normals
 
 
 @pytest.mark.parametrize('return_clipped', [True, False])
-def test_clip_filter(multiblock_all_with_nested_and_none, return_clipped):
+@pytest.mark.parametrize('crinkle', [True, False])
+def test_clip_filter(multiblock_all_with_nested_and_none, return_clipped, crinkle):
     """This tests the clip filter on all datatypes available filters"""
     # Remove None blocks in the root block but keep the none block in the nested MultiBlock
     multi = multiblock_all_with_nested_and_none
@@ -41,7 +42,9 @@ def test_clip_filter(multiblock_all_with_nested_and_none, return_clipped):
     assert None in multi.recursive_iterator()
 
     for dataset in multi:
-        clips = dataset.clip(normal='x', invert=True, return_clipped=return_clipped)
+        clips = dataset.clip(
+            normal='x', invert=True, return_clipped=return_clipped, crinkle=crinkle
+        )
         assert clips is not None
 
         if return_clipped:
@@ -60,6 +63,25 @@ def test_clip_filter(multiblock_all_with_nested_and_none, return_clipped):
                 assert clip.n_blocks == dataset.n_blocks
             else:
                 assert isinstance(clip, pv.UnstructuredGrid)
+
+
+@pytest.mark.parametrize('as_composite', [True, False])
+def test_clip_filter_pointset_no_points_removed(pointset, as_composite):
+    n_points_in = pointset.n_points
+    mesh = pv.MultiBlock([pointset]) if as_composite else pointset
+    # Make sure we clip such that none of the points are removed
+    bounds = pointset.bounds
+    clipped = mesh.clip(origin=(bounds.x_max + 1, bounds.y_max, bounds.z_max))
+    pointset_out = clipped[0] if as_composite else clipped
+
+    if as_composite and pv.vtk_version_info >= (9, 4) and pv.vtk_version_info < (9, 5):
+        assert pointset_out.is_empty
+        pytest.xfail("VTK 9.4 bug where clipping PointSet doesn't work")
+    assert np.allclose(clipped.bounds, bounds)
+
+    assert isinstance(pointset_out, pv.PointSet)
+    n_points_out = pointset_out.n_points
+    assert n_points_in == n_points_out
 
 
 def test_clip_filter_normal(datasets):
@@ -254,12 +276,8 @@ def test_extract_all_edges(datasets):
         assert edges is not None
         assert isinstance(edges, pv.PolyData)
 
-    if pv.vtk_version_info < (9, 1):
-        with pytest.raises(VTKVersionError):
-            datasets[0].extract_all_edges(use_all_points=True)
-    else:
-        edges = datasets[0].extract_all_edges(use_all_points=True)
-        assert edges.n_lines
+    edges = datasets[0].extract_all_edges(use_all_points=True)
+    assert edges.n_lines
 
 
 def test_extract_all_edges_no_data():
@@ -355,7 +373,6 @@ def test_cell_centers_no_cell_data(cube):
     assert not cube.cell_centers(pass_cell_data=False).cell_data
 
 
-@pytest.mark.needs_vtk_version(9, 1, 0)
 def test_cell_center_pointset(airplane):
     pointset = airplane.cast_to_pointset()
     result = pointset.cell_centers(progress_bar=True)
@@ -562,9 +579,6 @@ def test_cell_quality():
 def test_cell_quality_measures(ant):
     # Get quality measures from type hints
     hinted_measures = list(get_args(_CellQualityLiteral))
-    if pv.vtk_version_info < (9, 2):
-        # This measure was removed from VTK's API
-        hinted_measures.insert(1, 'aspect_beta')
 
     # Get quality measures from the VTK class
     actual_measures = list(_get_cell_quality_measures().keys())
