@@ -345,10 +345,9 @@ def test_disc_source():
     assert algo.outer == 0.5
     assert algo.r_res == 1
     assert algo.c_res == 6
-    if pv.vtk_version_info >= (9, 2):
-        center = (1.0, 2.0, 3.0)
-        algo = pv.DiscSource(center=center)
-        assert algo.center == center
+    center = (1.0, 2.0, 3.0)
+    algo = pv.DiscSource(center=center)
+    assert algo.center == center
 
 
 def test_cube_source():
@@ -376,9 +375,8 @@ def test_sphere_source():
     assert algo.start_phi == 0.0
     assert algo.end_phi == 180.0
     center = (1.0, 2.0, 3.0)
-    if pv.vtk_version_info >= (9, 2):
-        algo = pv.SphereSource(center=center)
-        assert algo.center == center
+    algo = pv.SphereSource(center=center)
+    assert algo.center == center
 
 
 def test_line_source():
